Implement maze.c
Try to solve the problem based on chessboard grid.
All the floor are initial with unvisited state. Starting from source A,
the area speards from A to achievable place. Since we spreads only to
the near grid, the spreaded place each round have the same color at
chessboard.

Because it is hard to know how many grid is spreaded, we compute whole
board at each round, using bitwise shift. If the grid is wall, it would
not be changed after the bitwise operations.