Fix infinite loop when cannot achieve
Modify the method we update matrix[].
The old version used |=. It causes the pioneer cannot identify 7 and 1
since they both have set bit at LSB. And this leads to infinite loop.
The new version uses | still, but introduing a bitmask to block the
target bits. If the number contains the target bits, it is updated by
left of |, on the other hand updated by right. A number would not
contain both. So, one effect of this commit is that we can no longer
know the position of source after beginning.